+++ RELEASENOTES.md
@@ -10,6 +10,7 @@ Release notes for pyraxshell
   * Lib/Plugin_Dns log to file/db (Plugin.r())
   * Lib/Plugin_Loadbalancers log to file/db (Plugin.r())
   * Lib/Plugin_Servers log to file/db (Plugin.r())
+  * Lib/Plugin_Services log to file/db (Plugin.r())
 
 ## 2013.09.29 - Version 0.2.4
   * Added SIGINT/SIGTERM handler to stop threads gracefully